+++ src/main/java/ru/itmo/aquacourseproject/repository/KingdomEntityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.itmo.aquacourseproject.entity.KingdomEntity;
 
 @Repository
-public interface KingdomEntityRepository extends CrudRepository<KingdomEntity, Long> {
+public interface KingdomEntityRepository extends CrudRepository<KingdomEntity, Integer> {
 }
