+++ src/main/java/ru/itmo/aquacourseproject/repository/WeaponEntityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.itmo.aquacourseproject.entity.WeaponEntity;
 
 @Repository
-public interface WeaponEntityRepository extends CrudRepository<WeaponEntity, Long> {
+public interface WeaponEntityRepository extends CrudRepository<WeaponEntity, Integer> {
 }
